BugFix:Missed frame bug
When Missed frame occurs, Metal-Hud shows it by double. (like 1.00)

I didn't find it because apple official website did not show this case.

When the missed frame occurs, it makes a bug. So, i fix it by change missedframe type int to double.

+++ MetalHudParser-SwiftUI/Models/DataSet.swift
@@ -6,7 +6,7 @@ class DataSet{
     var csvData : [[String]] = []
     
     var dataHeader: String = ""
-    var missedFrame: Int = 0
+    var missedFrame: Double = 0
     
     var gpuTimeData: [Double] = []
     var memoryData: [Double] = []
@@ -52,7 +52,7 @@ class DataSet{
                 case 0:
                     dataHeader = rawData[rowIndex][0]
                 case 1:
-                    missedFrame = missedFrame + Int(rawData[rowIndex][1])!
+                    missedFrame = missedFrame + Double(rawData[rowIndex][1])!
                 case 2:
                     memoryData.append(Double(rawData[rowIndex][2])!)
                 default:
